Don't allow an authorized user to delete himself. Fix minor display problem in changegroupassign
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/branches/1.2.x@5283 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ admin/listusers.php
@@ -169,7 +169,7 @@ if (FALSE == empty($error)) {
 				    {
 					echo "<td>&nbsp;</td>\n";
 					}
-				if ($remove && $oneuser->id != 1)
+				if ($remove && $oneuser->id != 1 && $oneuser->id != $userid)
 				    {
 					echo "<td><a href=\"deleteuser.php".$urlext."&amp;user_id=".$oneuser->id."\" onclick=\"return confirm('".cms_html_entity_decode_utf8(lang('deleteconfirm', $oneuser->username),true)."');\">";
                     echo $themeObject->DisplayImage('icons/system/delete.gif', lang('delete'),'','','systemicon');
